pyproject.toml
# pyproject.toml
[project]
name = "chatbot-rag"
version = "0.1.0"
description = "Chatbot con RAG, Streamlit y Ollama"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.46.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-chroma>=0.2.4",
    "langchain-ollama>=0.1.4", # Añadido para app.py
    "unstructured[pdf]>=0.18.2", # Usamos la versión más reciente con extras para PDF
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "loguru>=0.7.3",
    "pymupdf4llm>=0.0.25",
    "chonkie==1.0.8",
    "cache-to-disk>=2.0.0",
    "pdfminer.six>=20221105", # Usamos pdfminer.six que es la versión mantenida
    "easyocr>=1.7.2",
    "ipykernel>=6.29.5",
]

[tool.setuptools]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
    "loguru>=0.7.0",
    "tqdm>=4.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = ["notebooks", ".venv"]
extend-exclude = ["*.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]